+++ hurb/frontend-hurb-weather/src/components/BackgroundImage/index.ts
@@ -1,0 +1,2 @@
+export { BackgroundImage } from './BackgroundImage';
+export type { } from './BackgroundImage';