+++ hurb/frontend-hurb-weather/src/components/WeatherGrid/index.ts
@@ -1,0 +1,1 @@
+export { WeatherGrid } from './WeatherGrid';